Front End Client for the project
- NextJS Client
- Server side rendering and static generation for pages
- Client side rendering with react for interactive pages
- Implemented:
- Catalog view, filtering by make, type and history.
- Sorting by price, milage, asc and desc
- Add to shopping cart and remove from cart
- Redux store for storing user auth and shopping cart items
- Login, register and log out
- View reviews for cars

Catalog Microservice:
- Changed schema to include reviews and quantity

Shopping Cart Microservice:
- Changed schema to have quantity for each car

+++ shoppingCartServices/DAO.ts
@@ -1,56 +1,78 @@
 const mysql = require("mysql2/promise");
 import { Cart, PrismaClient } from "@prisma/client";
 import dotenv from "dotenv";
-import fs from "fs";
-import path from "path";
+
 dotenv.config();
 const prisma = new PrismaClient();
 
 export class DAO {
-  getAllCarts = async () => {
+  getAllCarts = async (): Promise<Cart[]> => {
     try {
-      const allCarts = await prisma.cart.findMany();
+      const allCarts = (await prisma.cart.findMany()) as Cart[];
       return allCarts;
     } catch (error: any) {
       return error;
     }
   };
 
-  getCartByID = async (id: number) => {
+  getCartByID = async (id: number): Promise<Cart[]> => {
     try {
-      const cart = (await prisma.cart.findMany({ where: { id: id } })) as Cart[];
+      const cart = await prisma.cart.findMany({ where: { id: id } });
       return cart;
     } catch (error: any) {
       return error;
     }
   };
 
-  addCart = async (cartId: number, carID: number) => {
+  addCart = async (cartID: number, carID: number): Promise<number> => {
     try {
-      const cart = await prisma.cart.create({ data: { id: cartId, carID: carID } });
-      return cart;
+      const previusQuantity = await this.getCarQuantity(cartID, carID);
+      if (previusQuantity) {
+        const cart = await prisma.cart.updateMany({
+          where: { id: cartID, carID: carID },
+          data: { quantity: previusQuantity + 1 },
+        });
+        return cart.count;
+      }
+      const cart = await prisma.cart.create({ data: { id: cartID, carID: carID, quantity: 1 } });
+      return cart.quantity;
     } catch (error: any) {
       return error;
     }
   };
 
-  async deleteCart(cartID: number) {
+  async deleteCart(cartID: number): Promise<number> {
     try {
       const cart = await prisma.cart.deleteMany({ where: { id: cartID } });
 
-      return cart;
+      return cart.count;
     } catch (error: any) {
       return error;
     }
   }
 
-  async deleteFromCart(cartID: number, carID: number) {
+  async deleteFromCart(cartID: number, carID: number): Promise<number> {
     try {
+      const previusQuantity = await this.getCarQuantity(cartID, carID);
+      if (previusQuantity && previusQuantity > 1) {
+        const cart = await prisma.cart.updateMany({
+          where: { id: cartID, carID: carID },
+          data: { quantity: previusQuantity - 1 },
+        });
+        return cart.count;
+      }
       const cart = await prisma.cart.deleteMany({ where: { id: cartID, carID: carID } });
-
-      return cart;
+      return cart.count;
     } catch (error: any) {
       return error;
+    }
+  }
+  async getCarQuantity(cartID: number, carID: number): Promise<number | null> {
+    try {
+      const previusQuantity = (await prisma.cart.findMany({ where: { id: cartID, carID: carID } }))[0].quantity;
+      return previusQuantity;
+    } catch (error) {
+      return null;
     }
   }
 }
